Update index.md
Removing this:

```

{copy for this file (index.md)}
Replace the text above the commented area with your information in the format below:
```


```
{info.md}

This separate file is where you should place links to your Google Group and Meetup page. It will be automatically rendered in the column sidebar.

{leaders.md}

Another separate file that should simply include each leaders name with mailto link as a list. It will also be automatically rendered in the column sidebar.

-->